Refactor to config, data, xfer packages
