+++ src/main/java/scholarshare/price/data/Fund.java
@@ -1,4 +1,4 @@
-package scholarshare.price;
+package scholarshare.price.data;
 
 import com.google.common.collect.Maps;
